pyproject.toml
[project]
name = "codingagent-mobile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.21.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
]
